+++ dao/src/main/java/by/mozgo/craps/dao/GameDao.java
@@ -16,5 +16,5 @@ public interface GameDao extends BaseDao<Game> {
      * @return number of games
      * @throws DaoException if a database access error occurs
      */
-    int findGamesNumber(int userId) throws DaoException;
+    int findGamesNumber(long userId) throws DaoException;
 }
